Add ability to set insecure via environment variable
When running a [custom container service] via cephadm it is not possible
to set any arguments for the container entrypoint. You can only set
arguments for the docker/podman command itself and set environment
variables. This works well with `RADOSGW_SERVER`, etc. but does not allow
for the usage of private certificates on the RGW endpoint when using the
exporter this way.

This commit adds the option to set `SECURE=false` to ignore certificate
errors (which I need – unfortunately).

This also adjusts the arguments name to use "secure" as a name to make
the code more readable.

[custom container service]: https://docs.ceph.com/en/quincy/cephadm/services/custom-container/